+++ src/web/mod.rs
@@ -1,3 +1,6 @@
 //! Modules and functions for running the web server.
+mod errors;
 pub mod handlers;
 pub mod routes;
+
+pub use errors::*;